refactor: unify embedding modules
