+++ changelog.d/2025.09.05.23.09.38.changed.md
@@ -1,0 +1,1 @@
+- Consolidated embedding modules into `@promethean/embedding` package and updated references.
