+++ docs/embedding.md
@@ -1,0 +1,4 @@
+# Embedding Package
+
+`@promethean/embedding` combines attachment and remote embedding utilities.
+It exposes an attachment embedder and a remote embedding function for broker-based embeddings.
