--- packages/attachment-embedder/ava.config.mjs
@@ -1,3 +1,0 @@
-export default {
-  files: ["dist/tests/**/*.js"],
-};
