--- packages/attachment-embedder/config/providers.yml
@@ -1,7 +1,0 @@
-providers:
-  - provider: discord
-    tenant: duck
-    credentials: {}
-    storage:
-      mongo_db: test
-      chroma_ns: discord__duck
