--- packages/attachment-embedder/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@promethean/attachment-embedder",
-  "version": "0.0.1",
-  "type": "module",
-  "main": "./src/index.ts",
-  "scripts": {
-    "start": "node --enable-source-maps --experimental-specifier-resolution=node ./src/index.ts",
-    "test": "pnpm build && ava --config ../../../config/ava.config.mjs tests/smoke.test.js",
-    "coverage": "pnpm build && c8 ava",
-    "build": "tsc",
-    "lint": "pnpm exec @biomejs/biome lint ."
-  },
-  "dependencies": {
-    "@promethean/migrations": "file:../migrations",
-    "@promethean/platform": "file:../platform",
-    "discord.js": "^14.17.3"
-  },
-  "devDependencies": {
-    "tsx": "^4.20.3"
-  }
-}
