--- packages/attachment-embedder/tests/smoke.test.js
@@ -1,5 +1,0 @@
-import test from "ava";
-
-test("placeholder", (t) => {
-  t.pass();
-});
