--- packages/attachment-embedder/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "../../config/tsconfig.service.json",
-  "compilerOptions": {
-    "outDir": "./dist",
-    "baseUrl": ".",
-    "rootDir": "./src",
-    "paths": {
-      "@shared/ts/*": ["../../../shared/ts/dist/*"]
-    }
-  },
-  "include": ["./src/**/*.ts"]
-}
