+++ packages/cephalon/package.json
@@ -31,7 +31,7 @@
     "@discordjs/opus": "^0.10.0",
     "@discordjs/voice": "^0.18.0",
     "@promethean/agent-ecs": "workspace:*",
-    "@promethean/embeddings": "file:../embeddings",
+    "@promethean/embedding": "workspace:*",
     "@promethean/legacy": "workspace:*",
     "@promethean/llm": "workspace:*",
     "@promethean/persistence": "workspace:*",
@@ -56,7 +56,7 @@
     "wav": "^1.0.2",
     "wav-decoder": "^1.3.0",
     "ws": "^8.17.0",
-    "@promethean/embeddings": "file:../embeddings",
+    "@promethean/embedding": "workspace:*",
     "@promethean/utils": "workspace:*",
     "@promethean/security": "workspace:*"
   },
