+++ packages/discord-embedder/tsconfig.json
@@ -10,8 +10,8 @@
       "@shared/js/*": ["../../js/*"],
       "@promethean/persistence": ["../persistence/dist/index"],
       "@promethean/persistence/*": ["../persistence/dist/*"],
-      "@promethean/embeddings": ["../embeddings/dist/index"],
-      "@promethean/embeddings/*": ["../embeddings/dist/*"]
+      "@promethean/embedding": ["../embedding/dist/index"],
+      "@promethean/embedding/*": ["../embedding/dist/*"]
     },
     "allowArbitraryExtensions": true,
     "types": ["node"]
