+++ packages/embedding/src/attachment/index.ts
