+++ packages/embedding/src/index.ts
@@ -1,0 +1,2 @@
+export { embedAttachments } from './attachment/index.js';
+export { RemoteEmbeddingFunction } from './remote.js';
