+++ packages/embedding/src/shims.d.ts
