+++ packages/embedding/src/tests/smoke.test.ts
@@ -1,0 +1,5 @@
+import test from 'ava';
+
+test('placeholder', t => {
+  t.pass();
+});
