--- packages/embeddings/.prettierrc.json
@@ -1,10 +1,0 @@
-{
-    "$schema": "https://json.schemastore.org/prettierrc.json",
-    "printWidth": 120,
-    "useTabs": false,
-    "tabWidth": 4,
-    "singleQuote": true,
-    "quoteProps": "as-needed",
-    "trailingComma": "all",
-    "endOfLine": "lf"
-}
