+++ packages/persistence/tsconfig.json
@@ -9,7 +9,7 @@
     "include": ["src/**/*"],
     "references": [
         {
-            "path": "../embeddings"
+            "path": "../embedding"
         }
     ]
 }
